feat: Add Python & Node.js bindings with Tiger Style polish
Add production-ready language bindings and complete remaining polish items.

**Python Bindings:**
- Clean ctypes wrapper with automatic memory management
- Type hints for all structures (_fields_: List[Tuple[str, Any]])
- Fixed bare except clause (now catches Exception explicitly)
- Input validation at FFI boundary (file size, params, etc.)
- Comprehensive tests and examples

**Node.js Bindings:**
- TypeScript-first design with full type safety
- Added pyjamaz_cleanup to FFI definitions
- Proper cleanup on process exit (SIGINT, SIGTERM, exit)
- Standardized error types (PyjamazBindingError for FFI layer)
- Both sync and async APIs
- Comprehensive tests (30+ total: TS + JS)

**Cache Safety:**
- Added bounds checking to parseMetadata in src/cache.zig
- 7 validation points (start < len, end > start) prevent panic
- Tiger Style: validate before slice operations

**Core Improvements:**
- Changed file_size from u32 to u64 throughout (EncodedCandidate, CacheMetadata)
- Refactored optimizeImage from 168 → 62 lines (≤70 line limit)
- All critical Tiger Style issues resolved
- FFI boundary documentation in src/CLAUDE.md

+++ src/test/conformance_runner.zig
@@ -28,7 +28,7 @@ const TestResult = struct {
     passed: bool,
     skipped: bool = false,
     input_bytes: u64,
-    output_bytes: ?u32,
+    output_bytes: ?u64, // Changed to u64 to match EncodedCandidate.file_size
     ratio: ?f64,
     diff_value: ?f64 = null, // Perceptual quality diff (0.0 = identical, higher = more different)
     reason: ?[]const u8,
